+++ frontend-set/src/components/common/Footer.vue
@@ -1,0 +1,24 @@
+<template>
+  <!-- 푸터 -->
+  <footer class="bg-gradient-to-r from-gray-100 to-gray-200 py-8 shadow-inner mt-16">
+    <div class="max-w-[1200px] mx-auto px-5">
+      <div class="flex justify-center space-x-8">
+        <a
+          href="#"
+          class="text-gray-600 hover:text-gray-900 cursor-pointer transition-colors hover:drop-shadow-sm"
+          >Terms and Conditions</a
+        >
+        <a
+          href="#"
+          class="text-gray-600 hover:text-gray-900 cursor-pointer transition-colors hover:drop-shadow-sm"
+          >Privacy Policy</a
+        >
+        <a
+          href="#"
+          class="text-gray-600 hover:text-gray-900 cursor-pointer transition-colors hover:drop-shadow-sm"
+          >Contact Us</a
+        >
+      </div>
+    </div>
+  </footer>
+</template>
